+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -16,6 +16,7 @@
 #include "Rule10p6Check.h"
 #include "Rule10p7Check.h"
 #include "Rule10p8Check.h"
+#include "Rule14p4Check.h"
 
 namespace clang::tidy {
 namespace misrac {
@@ -39,6 +40,8 @@ public:
         "misrac-rule-10p7");
     CheckFactories.registerCheck<Rule10p8Check>(
         "misrac-rule-10p8");
+    CheckFactories.registerCheck<Rule14p4Check>(
+        "misrac-rule-14p4");
   }
 };